Use Path to load and wirte token as OS-neutral
Path class from pathlib can find the file without having to worry
about / (Linux and MacOS) or \ (Windows).

It also provides methods to read and write files.

Since this moves to an OS-neutral way or writing and loading the token file,
this commits also changes the file name to just "key".